more efficient packing of config, layout and node structs, saved 36 bytes per node, 16 bytes per config and 16 bytes per layout results

+++ Krystal.UI/include/Krystal.UI/Algorithm/CalculateLayout.hpp
@@ -12,5 +12,5 @@ namespace Krys::UI
                                Direction ownerDirection, SizingMode widthSizingMode,
                                SizingMode heightSizingMode, float ownerWidth, float ownerHeight,
                                bool performLayout, LayoutPassReason reason, LayoutData &layoutMarkerData,
-                               uint32 depth, uint32 generationCount);
+                               uint32 depth, uint16 generationCount);
 }
